addressed deprecation warnings

+++ src/wakeexchange/OptimizationGroups.py
@@ -144,12 +144,13 @@ class OptAEP(Group):
         # ##### add constraint definitions
         self.add('spacing_con', ExecComp('sc = wtSeparationSquared-(minSpacing*rotorDiameter[0])**2',
                                          minSpacing=minSpacing, rotorDiameter=np.zeros(nTurbines),
-                                         sc=np.zeros(((nTurbines-1.)*nTurbines/2.)),
-                                         wtSeparationSquared=np.zeros(((nTurbines-1.)*nTurbines/2.))),
+                                         sc=np.zeros(int(((nTurbines-1.)*nTurbines/2.))),
+                                         wtSeparationSquared=np.zeros(int(((nTurbines-1.)*nTurbines/2.)))),
                  promotes=['*'])
 
         # add objective component
         self.add('obj_comp', ExecComp('obj = -1.*AEP', AEP=0.0), promotes=['*'])
+
 
 class OptCOE(Group):
     """
@@ -267,8 +268,8 @@ class OptCOE(Group):
         # ##### add constraint definitions
         self.add('spacing_con', ExecComp('sc = wtSeparationSquared-(minSpacing*rotorDiameter[0])**2',
                                          minSpacing=minSpacing, rotorDiameter=np.zeros(nTurbines),
-                                         sc=np.zeros(((nTurbines-1.)*nTurbines/2.)),
-                                         wtSeparationSquared=np.zeros(((nTurbines-1.)*nTurbines/2.))),
+                                         sc=np.zeros(int(((nTurbines-1.)*nTurbines/2.))),
+                                         wtSeparationSquared=np.zeros(int(((nTurbines-1.)*nTurbines/2.)))),
                  promotes=['*'])
 
         # add objective component
